Fix memory leaks: 1. Remove secondary sentinel when checkpoint is running. 2. Overwrite deleted but not removed rows, due to checpoint or concurrent inserts reclaim old version using GC.

+++ src/gausskernel/storage/mot/core/src/system/transaction_logger/redo_log.cpp
@@ -370,9 +370,11 @@ RC RedoLog::SerializeTransaction()
                 case INS:
                     if (access->m_params.IsPrimarySentinel()) {
                         if (access->m_params.IsUpgradeInsert()) {
-                            status = DeleteRow(access->m_localRow);
-                            if (status != RC_OK) {
-                                return status;
+                            if (access->m_params.IsDummyDeletedRow() == false) {
+                                status = DeleteRow(access->m_localRow);
+                                if (status != RC_OK) {
+                                    return status;
+                                }
                             }
                         }
                         status = InsertRow(access->GetTxnRow());
